+++ packages/localstorage-adapter/src/adapter.ts
