--- packages/localstorage-adapter/src/adapter/index.ts
@@ -1,4 +1,0 @@
-const version = '__@FLOPFLIP/VERSION_OF_RELEASE__';
-
-export { version };
-export { adapter, STORAGE_SLICE } from './adapter';
